fix(ui):moved tags component
