+++ src/sentry/static/sentry/app/components/deployBadge.tsx
@@ -4,7 +4,7 @@ import styled from '@emotion/styled';
 import {t} from 'app/locale';
 import space from 'app/styles/space';
 import {Deploy} from 'app/types';
-import Tag from 'app/views/settings/components/tag';
+import Tag from 'app/components/tag';
 import Link from 'app/components/links/link';
 import {IconOpen} from 'app/icons';
 import {stringifyQueryObject, QueryResults} from 'app/utils/tokenizeSearch';
