+++ src/sentry/static/sentry/app/views/settings/organizationAuth/providerItem.jsx
@@ -12,7 +12,7 @@ import FeatureDisabled from 'app/components/acl/featureDisabled';
 import Hovercard from 'app/components/hovercard';
 import SentryTypes from 'app/sentryTypes';
 import {IconLock} from 'app/icons';
-import Tag from 'app/views/settings/components/tag';
+import Tag from 'app/components/tag';
 import {descopeFeatureName} from 'app/utils';
 
 export default class ProviderItem extends React.PureComponent {
