+++ src/sentry/static/sentry/app/views/settings/projectSourceMaps/detail/sourceMapsArtifactRow.tsx
@@ -12,8 +12,7 @@ import {Artifact} from 'app/types';
 import Confirm from 'app/components/confirm';
 import Access from 'app/components/acl/access';
 import Tooltip from 'app/components/tooltip';
-
-import Tag from '../../components/tag';
+import Tag from 'app/components/tag';
 
 type Props = {
   artifact: Artifact;
